requirements.txt
python-binance>=1.0.17
pytest>=7.0
